+++ NEWS.md
@@ -1,6 +1,9 @@
 # iraceplot 1.01
 
  * New function `ablation_plot()` to plot the result of `irace::ablation()`.
+
+ * New function `summarise_by_configuration()`.
+ 
  
 # iraceplot 1.0